+++ src/mod/AutoTextHeight.vue
@@ -43,7 +43,7 @@ export default {
   computed: {
     tipOption () {
       let option
-      if (this.isEllipsis) {
+      if (this.isEllipsis || this.tip === false) {
         if (typeof this.tip == 'object') {
           option = this.tip
         } else {
@@ -51,7 +51,7 @@ export default {
             placement: this.tip || 'top'
           }
         }
-        if (!option.title) {
+        if (option.title === undefined) {
           option.title = this.text
         }
       } else {